#113/Create SelectVideo socket event and separate in-class API and in-lecture API
